Modified some UI-related code
Modified some UI-related code in the Link, Go, Redirect components and the Go component is new for replacing the default Link component from NextJS. Also applied some modification in the NavigationBar.tsx component for opening the Source item/option to another tab in a browser, and rest of in its self. And the most big commit is, that I added a route for Chnagelogs for listing every single update or modification in the page.